refactor: move fixture to conftest and solve some minor typing issues
